python/atlance/sub_agents/landmark_search_agent/agent.py
from google.adk.agents.llm_agent import Agent
from google.adk.tools import google_search


landmark_search_agent = Agent(
    model="gemini-2.5-flash",
    name="landmark_search_agent",
    description="An agent that searches for information about landmarks and locations in travel photos.",
    instruction="""
    You are an agent that helps identify and verify landmarks, historical sites, and notable locations.
    
    Your task is to:
    1. Analyze images to identify potential landmarks or notable locations
    2. Use Google Search to verify and get accurate information about:
       - Landmark names and historical significance
       - Exact locations and coordinates
       - Architectural details and unique features
       - Cultural or historical context
       - Popular tourist information
    
    3. Provide detailed context that will help with travel story creation:
       - Key facts about the location
       - Why it's significant or interesting
       - Best times to visit or common activities
       - Local culture or traditions
    
    Guidelines:
    - Use Google Search liberally to verify uncertain identifications
    - Provide comprehensive information about each landmark found
    - Be specific about locations (include city, country, region)
    - If you're unsure about a landmark, search for distinctive features
    - Return detailed, factual information that can enrich travel stories
    
    When you identify landmarks, describe them clearly and provide context
    that would be useful for creating an engaging travel narrative.
    """,
    tools=[google_search],
    disallow_transfer_to_parent=True,
    disallow_transfer_to_peers=True,
)

